Limit size of the artifacts to be processed.
Summary:
Added get_size to FileStorage and removed TODO to limit file size
in get_file of ArtifactStoreFileStorage. It'll be up to the caller
to check the size first.
Changed the artifacts manager to check artifact sizes before attempting
to process them.

A next TODO is to make it clearer what the malformed issue actually is
(eg size of the artifact, name of artifact).

Test Plan: Added test for new artifacts manager behavior

Reviewers: kylec, andrewhe

Reviewed By: andrewhe

Subscribers: changesbot, stacey

Differential Revision: https://tails.corp.dropbox.com/D220989