Grouper integration with user admin status
Summary:
Historically, making/revoking an admin requires going through the Changes UI. We want to do this on Grouper so that we can say things like, "All members of Team A should be an admin".

This diff creates a recurring task every 1 minute that connects to Grouper, gets the list of admin users, and makes sure that those users and only those users have admin access.

This also updates the web UI to remove the "Make Admin" button from the users page.

The API endpoint that can manipulate the admin status of the user is purposely kept for compatability reasons. It is not a serious risk, as the Grouper sync will override any changes every 1 minute.

Test Plan:
unit test, manual testing on the VM

screenshot of UI change:
{F492038}

Reviewers: anupc

Reviewed By: anupc

Subscribers: changesbot, herb, kylec

Tags: #changes_ui

Differential Revision: https://tails.corp.dropbox.com/D219947