Remove model change for grouper sync
Summary: The issue with the diff was that we implemented the migration change and the model change in the same commit. This removes the model change. It also changes the migration code to make the column nullable for compatibility reasons (it is USUALLY a bad idea to edit the migration file, but since we haven't applied this anywhere, it should be fine)

Test Plan: tested locally, ran tests

Reviewers: anupc

Reviewed By: anupc

Subscribers: changesbot, kylec

Differential Revision: https://tails.corp.dropbox.com/D221771