Add status argument to jobstep_aggregate_by_status
Summary: jobstep_aggregate_by_status can sometimes be slow (2-3 seconds per request), since we're loading thousands of items from the database. In some cases, like with autoscaler, we only want to get numbers for a particular type of status, such as just counts for jobsteps that are currently `pending_allocation`. This change lets us specify a specific `status` to get stats on, which significantly speeds up these kinds of requests.

Test Plan: Added a unit test for this new functionality.

Reviewers: kylec

Reviewed By: kylec

Subscribers: changesbot, treaster

Differential Revision: https://tails.corp.dropbox.com/D222177