Updated sync_grouper to not needlessly query the database when given an empty set of emails
Summary: This is in response to warnings like https://forge-sentry.pp.dropbox.com/sentry/changes/issues/79/. When we are given an empty set of admin emails, or an empty dictionary of project admins, we still do a query of (~User.email.in_(the empty list)), which is trivially true and expensive. This diff avoids doing that.

Test Plan: unit tests

Reviewers: kylec, anupc

Reviewed By: anupc

Subscribers: changesbot, herb, kylec

Differential Revision: https://tails.corp.dropbox.com/D223113