Have collect target script return the artifact search path, and use that information
Summary: IN the past, we specified a static artifact search path for every kind of job steps. The problem with this is that with Bazel, the path is not static, but (at the time of this writing) the hash of the workspace path. The solution is, during the collect targets step, also return the artifact search path. This diff does that and also propagate the returned path to the expanded job steps.

Test Plan: unit tests

Reviewers: anupc

Reviewed By: anupc

Subscribers: changesbot, kylec

Differential Revision: https://tails.corp.dropbox.com/D224447