Add exception reporting to ISODatetime
Summary:
We see exceptions from this validator in prod, but we only record the error message
we return, which deprives us of the details of the real exception.
This adds exception logging when we see failures so that we can see what is really happening.
The added passing test case is a value that appears to fail in the wild.

Test Plan: A new test.

Reviewers: naphat

Reviewed By: naphat

Subscribers: changesbot, anupc

Differential Revision: https://tails.corp.dropbox.com/D225616