Remove the 'changed tests' info from the build details API
Summary:
Data isn't exposed in the UI or used by any API clients I'm aware of, and at a spot-check wasn't
correct. But even if it were correct, it's taking up >1/3 of the build api endpoint for data we
aren't using, so seems best to remove it and look to our history if we want to revive.

Test Plan: Changes

Reviewers: naphat

Reviewed By: naphat

Subscribers: changesbot, anupc

Differential Revision: https://tails.corp.dropbox.com/D227879