Route delete tasks to their own queue.
Summary:
The current deletion tasks can take a long time and they can potentially
block other tasks. This makes it so that the tasks go to their own queue
that can be processed by separate workers.

If workers are configured to read from all queues, this should be a noop.
Otherwise, new workers are needed that will read from this particular queue.

Reviewers: kylec

Reviewed By: kylec

Subscribers: changesbot, anupc

Differential Revision: https://tails.corp.dropbox.com/D231714