Only test targets that are affected by change
Summary:
This diff implements a part of our selective testing picture. If `selective_testing_policy` is set to `enabled` for a build, only affected targets for that build will be tested. The other targets will be created, but marked as unaffected `ResultSource.from_parent`).

Note that this diff does not enable selective testing - the variable `selective_testing_policy` on the build object is never set and defaults to `disabled`.

This diff also does not do any result propagation / revision result creation.

Test Plan: unit tests

Reviewers: anupc

Reviewed By: anupc

Subscribers: changesbot, kylec, wwu

Differential Revision: https://tails.corp.dropbox.com/D229863