Limit test results provided to email templates
Summary:
This is mostly a self-defense measure, to keep from potentially processing
thousands and thousands of tests.
If users want more than a handful of tests, they can go to the UI.
Since we also report the test count, the risk of users assuming they're seeing
the complete failed picture is reduced.

Test Plan: Existing

Reviewers: paulruan

Reviewed By: paulruan

Subscribers: anupc

Differential Revision: https://tails.corp.dropbox.com/D232011