Add index to Task.status
Summary:
We have monitoring and clean-up tasks that engage in table scans because they primarily care about
status; this is currently triggering our long-query warnings and creating noise.
Hopefully an index fixes that.
Size of the table is ~3m items in prod, so the indexing may take a bit, but we don't need
the index immediately, so adding it concurrently seems like a way to avoid downtime.

Test Plan: Staging, potentially a (new?) read replica

Reviewers: naphat

Reviewed By: naphat

Subscribers: changesbot, anupc

Differential Revision: https://tails.corp.dropbox.com/D233244