Add index to Task.status
Summary:
We have monitoring and clean-up tasks that engage in table scans because they primarily care about
status; this is currently triggering our long-query warnings and creating noise.
Hopefully an index fixes that.
This is the same as 8c49ac5821ad7af6dce16022501c992278c5ac97 but not concurrent.

Test Plan: None

Reviewers: naphat

Reviewed By: naphat

Subscribers: changesbot, anupc

Differential Revision: https://tails.corp.dropbox.com/D233891