Limit time frame of tests to delete per task and run the clean up task more frequently.
Summary:
This is to avoid the issue of having very long running delete queries.

Added a chunk size parameter to clean_project_tests that controls how big of a time frame
we look for tests to delete.
Replaced the delete-old-data task with two tasks: one that only deletes 10m worth of tests
and another that looks at 5hrs but with a time frame that should've been cleaned by the
quicker task.

Because all clean up tasks are bounded in how far back to look now, very old tests will have
to be manually cleaned up.

Test Plan:
Added a test for bounding the time frame in clean_project_tests.
Tried running a delete for a 10m time frame (>200k rows) on a no traffic replica db and
it took more than 6min.

Reviewers: kylec

Reviewed By: kylec

Subscribers: changesbot, anupc, kylec

Differential Revision: https://tails.corp.dropbox.com/D234367