only allow selective testing if the project is an autobazel project. enable selective testing for phabricator diff builds
Summary:
This updates build_index and build_retry to check if the project is an autobazel project before allowing selective testing (previously, the selective testing policy would not be set to enabled, but a message will be displayed on the build saying that `bazel.selective_testing_enabled` is false, which is only relevant for autobazel projects).

This also enables selective testing for phabricator diff builds. The default behavior for phabricator diff builds is to enable selective testing.

Test Plan: unit tests

Reviewers: anupc

Reviewed By: anupc

Subscribers: changesbot, kylec, wwu

Differential Revision: https://tails.corp.dropbox.com/D235521